tinkering

+++ zzflate/zzflate.h
@@ -1,10 +1,8 @@
 #ifndef _ZZFLATE
 #define _ZZFLATE
 
-
 #include <stdint.h>
 #include <memory>
- 
 #include <functional>
   
 enum Format {Zlib, Gzip, Deflate};
@@ -20,7 +18,7 @@ struct Config
  
 void StaticInit();
  
-void ZzFlateEncode(uint8_t *dest, unsigned long *destLen, const uint8_t *source, size_t sourceLen, const Config* config);
+void ZzFlateEncode(uint8_t *dest, size_t*destLen, const uint8_t *source, size_t sourceLen, const Config* config);
  
 void ZzFlateEncodeToCallback(const uint8_t *source, size_t sourceLen, const Config* config, std::function<bool(const uint8_t*, size_t)> callback);
